Added the files
The index.js module introduces three methods, the main method, which is the entry point of the application. In the main() method the openAi programming interface is contacted, to prompt the AI model and to generate content. The content is generated on two stages, first it generates ideas, then it generates script according to the ideas. The index has two more methods writeToFile() and sanitizeInput(). The write to file method exports the generated script ideas content to a file.
The visialsGenerator module has a purpose of extracting the visual part of the script and passing it to the model generating videos. It has method getFileContent which reads the content from the file the index module created. It has a method to sanitize the input, removing all special characters that might occure in it and finally it has a method to extract just the visuals part of it.