New prompts and new architecture
Changed the way the visualsGenerator module gets the information. Instead of reading from the file containing the script, the content is being directly passed to the visualsGenerator from the index module. This way I promote faster execution time, as there are less async operations taking place.
Also changed the way the methods are encapsulated and involked in the same visualsGenerator model.  The method visualsGenerator now involves all of the other different methods for sanitizing the raw data and also extracting the visuals of it.  As the visualsGenerator model does not read from a file, the readFileContent() method has been removed.
The prompts for prompting the chatGPT4 APIs has been re-crafted and improved, to encourage a maximum quality of the generated final product.