Re-wrote the application according to the MVC structural pattern
Re-organised the application to implement the MVC structural pattern architecture.
Re-wrote all of the different modules into classes, to live up to the OOP standards and guidelines.
The GPT model family is being extended from a GPTModel class, that encaplsulates the logic for instantiating the OpenAI - chatGPT API access point.
The application introduces two main classes for generating textual content with the use of OpenAI -chatgpt APIs. There is the GPTScriptGenerator, that generates a more generic textual content, based on a pre-set prompt message. Then there is the GPTScriptGeneratorCustomizable, that generates script, based on parameters passed as an object to it.
There is also a ContentExtractor class that has two methods, one for extracting the script content and one for extracting the music keywords
Another important class is introduced, the ContentSanitizer class, which sanitizes the textual output from the content generating moduls