Implemented LangChain integration, merged to mjs
Translated all of the normal js modules into mjs modules, translating imports and exports.
Finished the integration of LangChain into the pipeline. In the particular case, lang chain provides the possibility to chain custom textual data to the generic capabilities of the OpenAI LLM, gpt-4.
The implementation consists of two magor parts, the DataFedModel class, that encapsulates the logic for chaining the LLM, together with the textual content split into smaller data chunks, and invoking it into a chain. The DataFedModel class consists of one publicly accessible method for generating content based on parameters passed to it. It requires a system message parameter and user message parameter, combined together they compose a whole prompt.
The second part of the implementation  is the GPTScriptGeneratorDataFed, which is a copy of the GPTScriptGeneratorCustomizable, but with the difference, behind the scenes it uses the DataFedModel model.