New models architecture Text-to-speech generator
Changed the architecture of the Fliki.ai model. Ecapsulated the callApi method, which is the entry point to the API in a class called "VideoModel", the name suggests it contacts the LM that generates our video content, the same is also used to generate text-to-speech content.  The video generator class, now extends the VideoModel. This was done, to simplify  the work load later.
Composed a new class, that is a text-to-speech generator, called SpeechGenerator, the class extends the VideoMaker, as the text-to-speech service is also provided by fliki.ai